fix: BaseEntity id 타입 변경(int → long)에 따른 Repository 제네릭 타입 반영

+++ back/src/main/java/com/back/domain/party/party/repository/PartyRepository.java
@@ -3,5 +3,5 @@ package com.back.domain.party.party.repository;
 import com.back.domain.party.party.entity.Party;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartyRepository extends JpaRepository<Party, Integer> {
+public interface PartyRepository extends JpaRepository<Party, Long> {
 }
